backend/requirements.txt
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
billiard==4.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
httptools==0.6.0
idna==3.4
kombu==5.3.2
prompt-toolkit==3.0.39
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.23.2
uvloop==0.17.0
vine==5.0.0
watchfiles==0.20.0
wcwidth==0.2.6

# Websocket
websockets==11.0.3
websocket-client==1.6.3

# SqlAlchemy
async-sqlalchemy==1.0.0
alembic==1.12.0
asyncpg==0.28.0
psycopg2-binary==2.9.7
psycopg2==2.9.7

# Binance
python-binance==1.0.19

# Redis
aioredis==2.0.1
redis==5.0.0

#Celery
flower==2.0.1
celery==5.3.4
